acq: implement module.
This module is the most complex one so far. We use a separate thread for
handling acquisitions and avoiding blocking in unsightly ways any other
PV handling. There's still a decision to be taken on how parameter
changes are supposed to affect acquisitions that are already in effect;
for now, they are taken into account for the next acquisition. Stopping
an acquisition has yet to be implemented, as well.

We also use readEnum generously, in order to have all the enum
information in the driver.