iocutca: instantiate acq module.
For this commit, the changes weren't restricted to the module-specific
.cmd files, since we had to add new environment variables, and it made
more sense to declare the data in rtmlamp.substitutions. For this
reason, we re-ordered file sourcing in st.cmd, since acq.cmd had to be
sourced first.

It was necessary to fix rtmlamp.substitutions as well, since it was
missing the last character of each of the last four channels' names.

raw_data.template using NELM isn't ideal, since the data can contain
many atoms, and its max number of elements should be number_of_atoms
times max_desired_samples.

We also use a different ACQ_NAME here from what's currently deployed, in
order to describe what the acquisition module is actually doing.