iocutca: add runFOFB.sh script.
This script is used to launch a FOFB IOC for a physical slot that's
passed as a command line argument. It uses the new getCrate.sh and
getSlot.sh scripts; the former uses the machine hostname to determine
what crate it belongs to, and the latter translates a physical slot
(including the possible "slot-1" suffix) into a virtual slot (we are
still using virtual slots, to avoid having to change fofb-slot-mapping,
which was copied directly from fofb-epics-ioc).

For the procServ invocation, we'd like to use the --oneshot option, but
it's not yet available in the version we have installed.