acq: fix stop command for repetitive acquisitions.
If a stop command was sent, acquisitions would be stopped momentarily,
but then the start command from the repetitive logic would be read from
the queue again and acquisitions would be restarted.

Now, if there are any commands in the queue, we don't send a start
command for repetitive acquisitions.