rtmlamp: add RB records for gains and offsets.
Having only SP records breaks the API expected by most of our utilities,
and for the case of current offset, is forward compatible with our idea
of supporting it as a hardware register.

It was necessary to use PP in the OUT links, otherwise the -RB records
weren't processed (didn't generate monitor events when changed, which is
an issue with visualizer interfaces, for example).