fofb_processing: fix monitor client overload.
asyn's doCallbacks*Array() functions don't check if there have been any
changes to the underlying array, and neither does the EPICS DB layer, so
every call to these functions was generating a monitor event for CA
clients. Since they were being called in read_parameters(), we were
overloading applications that connected to the RefOrb and InvRespMatRow
PVs, e.g. the high-level FOFB IOC.

Since we are the only ones in charge of the hardware, and these values
only need to be read back from the hardware when there's a write, we can
call the callbacks only in the write*Array() functions.

A better design for read_parameters() could only touch the parameters
that actually change by themselves, and a separate version of the
function would be called after parameter writes. However, this would
still lead to spurious monitor events for unrelated PVs after writes, so
the future change will probably be layered on top (or below, seeing as
part of it will be in μHAL) of the current solution.