acq: implement data splitting for the sysid core.
Add parameters for all the resulting vectors as well.

This development made us realise the acq controller should expose the
number of atoms in the result value, so it doesn't have to be hardcoded
in the IOC. This allows the sample size to change to fit new data, and,
as long as the atom positions are kept, the code doesn't have to be
updated.

Since the sysid core can also be characterized as a FOFB acquisition,
the channel used for DCC acquisitions has been renamed.

The fofb-ctrl-gw commit with the final version of the acquisition
definition [1] was used as reference. The definition there goes from
most to least significant bits, while we generally read them from least
to most significant. Due to our little endianness (this is expected to
run on x86_64), all 16-bit atoms switch places inside a 32-bit double
word, which is why the 16-bit atoms aren't ordered strictly from N to 0
in the gateware (this is so the least significant half of the double
word contains the second value, and the most significant half contains
the first one - we read the acquisition as a whole from LSB to MSB, but
inside the double word we treat its most significant word first).

[1] https://github.com/lnls-dig/fofb-ctrl-gw/commit/f12460f26796a1a98fc243c85dc27fdd9d72d5f4