fofb_cc: don't use parameter library for writes.
It can get confusing to use the parameter library for dealing with
parameter writes, since these values are updated after writes, when
read_parameters() is called, and there can be weird divergences. We
should always write directly into the ctl structure, which is what's
being done for most other modules. The only one that still needs to
transition to this paradigm is the rtmlamp driver.