service: add files for FOFB IOC.
These scripts and unit files expect the erics directory to be in
/opt/erics, and for a bpm-epics-ioc user to be available (this was done
to take advantage of the existing user on our crates).

MemoryAccounting is left enabled, for debugging and provisioning
purposes. MemoryLimit is set as an experiment, in order to increase
reliability of the whole OS at the (possible) expense of this specific
service; we should note that memory usage seems to sit below 100MB at
most times, but there hasn't been any long term testing.