fofb_processing: add autosave config.
FOFBAccGain-SP is left off of the config, since it's better to have it
as bumpless reboot: it is changed often and there isn't a good
"start-up" value for it.

Also move PINI field around for consistency, and fix SCAN for
LoopIntlk-Mon (it was working "correctly", because of the scan task and
the defaut implementation for readInt32, but it's cleaner this way).