asub-rountines: implement asub_goff.
The best way of handling converting data from arrays without depending
on another module (calc) and with the option of propagating NORD from
one array to another seems to be by using aSubRecord.

Our asub_goff implementation is generic, so that it can be used for
cases when the offset should be applied both before and after
multiplication by a gain. The implementation choice of two offsets
rather than two gains was due to efficiency: it is expected that
addition will be faster than multiplication.

We originally passed NORD from the input array as an additional input to
aSub by itself, but turns out it isn't necessary [1], because NEA is set
automatically.

Should we switch to C++20, our template declarations can be simplified
by using auto in the parameters [2].

[1] https://epics.anl.gov/tech-talk/2023/msg00634.php
[2] https://stackoverflow.com/a/60355539