rtmlamp: fix unresponsive state in Current-SP.
After a transition from "fofb" to "manual" in OpMode-Sel happened, we
observed that the Current-SP record wouldn't process when a new value
was first written to it (so the new current wouldn't be written to
hardware via CurrentRaw-SP), nor would it be processed if one tried to
write that same value again; it was necessary to write a different value
altogether to force it to be processed.

This issue seems to have been caused by the logic which propagates the
value written in CurrentRaw-SP to Current-SP. To avoid complicating any
such logic even further, we have decided that the possibility of a small
precision error happening when converting the ref current to amperes is
not a big problem (the error would be small, and definitely below the
noise threshold of the ADC anyway), so now we write through Current-SP
directly, and there's no need to deal with complicated record logic.