afc_timing: improve FPGAClk-Cte initialization.
Make FPGAClk-Cte an ao record so we can use OMSL "closed_loop" to read
the EVG's FPGAClk, which means we can now add FPGAClk-Cte.DOL to
autosave as well, and make it configurable (this wasn't possible with
INP, because it was going to be erased by DisableFPGAClk).
DisableFPGAClk can be much simpler as well, since it only needs to move
FPGAClk-Cte.OMSL to "supervisory"; this means any users who want to
force a resync can simply set FPGAClk-Cte.OMSL to "closed_loop".

We also add FPGAClk-Cte to autosave, so any spurious network failures
won't stop us from bringing up the system or require manual intervention
to set FPGAClk-Cte. However, we can't add PINI=YES to FPGAClk-Cte, since
it would move it to "supervisory" mode, which means we need to add
PINI=YES to the main records that use it.