acq: add proper -RBs for some PVs.
We use -RB PVs to indicate that an IOC has processed and accepted a
given setpoint, so using DB links directly is not an issue. These
changes don't cover all of the acq PVs, but they cover the most relevant
ones in use with these IOCs. The readback values from hardware are now
-Mon PVs.

Also clean up extraneous VAL from SamplesPost-RB.