fofb_processing: add gain for fofb_shaper_filt.
The filter defined in fofb_shaper_filt has representation limits for its
coefficients, so an external multiplier gain might be necessary to
actually implement it in hardware. To avoid requiring an additional gain
in the hardware and the resultant resource consumption, it was decided
to use the FOFB accumulator gain implemented in hardware as the
combination of FOFBAccGain and FOFBAccFilterGain.

It is simpler to define the PV for it in fofb_processing, since it has
to be combined with the FOFBAccGain for the FOFBAccHwGain that's applied
to the hardware.

The output for FOFBAccHwGain is a -RB because it shouldn't be written
to, and is the configuration we tried to write to hardware. The input is
a -Mon because its sole purpose is getting a feedback as to what's
actually in hardware.

In order to provide feedback on the applied gain's precision, the
readback for the separate gains uses the gain that was actually applied
to hardware divided by the other gain's setpoint.

The slightly complex CALC definition for the gain readbacks is necessary
for two reasons: the readback will be different from the setpoint if the
resulting HwGain is outside of the hardware's representable range (this
can also be verified with FOFBAccHwGain-Mon's value); and one of the
gains being set to 0 shouldn't cause the readbacks to be NaN.