udriver: add member for read-only params.
Instead of having to define p_* members in subclasses for all the
read-only parameters, they can simply refer to p_read_only, minimizing
the amount of necessary declarations in class definitions.

It is necessary to declare it as a static class member to avoid a "base
class is uninitialized when used here to access" warning; and as an
inline class member so it doesn't have to be defined outside of the
class declaration.

Out of an abundance of caution for how this might work in a
multi-threaded scenario (which shouldn't be the case for us at the
moment), we avoid actually passing p_read_only to
asynPortDriver::createParam() and use a temporary integer.