utcaApp: add Interlock ASG.
This ASG will be used by the orbit interlock module on BPMs.

To avoid errors during asInit, we add an asSetSubstitutions call to
common.cmd. Since run.sh for BPMs doesn't define the environment
varibles for P and R in common.cmd, we need an additional call in
stBPM.cmd.