utcaApp: use fill_enum.
Changing the acq module over to use this function is left for later,
since it has more complex logic across its enums, including usage of
severity.

It doesn't make sense to spend lines of code setting *nIn to zero when
we return asynError and the resulting vectors won't be used, so we
remove those.