udriver: add parameter property vector to class.
This is a cleaner and more performant way of tracking parameter
properties (such as write-only and type) than adding sets for each
boolean property (which also means properties with more than one
possibility weren't covered adequately).

As a bonus, this improves the interface for defining these properties,
which can be done directly with the parameter list passed to
UDriver::UDriver().