pos_calc: add records for position calculation.
The aSub record will be triggered by the FIFO event and will process the
ADC Ampl records to update their value, via the input links with PP.

Since there isn't a guarantee that the position will have changed value
every time it's processed (the antenna values might be all the same),
the records are instantiated with MDEL=-1, so that they always generate
monitor events for orbit consumers. The amplitude records are also
changed to MDEL=-1.