acq: define data_type in initialization.
Having the data type as a PV which can be set at runtime was used mainly
for debugging and as a workaround for a limitation in the standard
UDriver iocsh function definition. For production, it only adds an
additional failure mode, and complicates support for different board
types.

An additional advantage is being able to instantiate only the relevant
result parameters.

This commit fixes a segfault in the BPM ACQ cores caused by commit
cda0b61457c1d05c7554c09f022bb063dc6279bb. It didn't take BPMs into
account, since this IOC hasn't been deployed for them yet.

It was necessary to include changes to the driver source, records and
.cmd files in the same commit, since they are all affected.