si57x_ctrl: implement module.
This module should have the ASYN_CANBLOCK flag, since its implementation
library blocks when accessing the device. This is not critical, however,
since the main time when it receives write requests is during iocInit.

We need to handle read_startup_regs() specially, since some AFC BPM
boards don't have both FMCs installed, so we can't throw an exception
and abort the IOC in case initializing the missing device fails. This
doesn't cover the case when the core never stops being busy, where
initialization simply stops, but it at least allows the IOC to startup
when that doesn't happen.