si57x_ctrl: add template file.
It was necessary to add an alias for INFOClkFreq, because the old IOC
had a mechanism to enable ADCSi57xFreq to follow the value of
INFOClkFreq. Since this mechanism isn't necessary for us, we can simply
omit it, and just make both PVs available for legacy clients, which
depend on the value of INFOClkFreq.