asub-routines: implement polynomial correction for pos_calc.
This polynomial correction is reponsible for applying the non-linear
correction to the beam's X, Y, Sum and Q values. Differently from the
correction applied in bpm-epics-ioc, which requires that the polynomial
coefficients include the gains in order to reach the correct magnitude,
we use dimensionless coefficients, and apply the gain and offset as a
last step.