service: start RFFE IOC for BPMs.
Whenever a BPM board is detected, both the AFC and RFFE IOCs should be
started automatically. Since RFFEs are connected to a single BPM, it is
necessary to start up two RFFE IOCs for each BPM board. In order to
track these services, their virtual slot is used. Physical slot 12 is an
exceptional case as it's always only connected to one BPM, so we don't
run the RFFE IOC for virtual slot 24.

So the script doesn't require the RFFE IOC being installed on the
system, making it optional to successfully start up its services.