fmc_active_clk: add ClksLocked record.
Looking at only the PLL or MMCM clock status is not enough to fully
represent the status of the BPM clock chain.

We implement this record in fmc_active_clk because the MMCM clock always
exists, but the AD9510 PLL clock is only available if the fmc_active_clk
module is present.