fmc250m: add template file.
Using this module, we implement a feature to automatically reset the
relevant devices (ISLA216p ADCs and FPGA MMCM) when there is a
transition from the device clocks being "not_locked" to "locked". This
situation can happen during board startup, when the ADCs receive no
clock signal, and can only be properly reset after they start receiving
it; and during operation, if the clock reference is lost, making it
necessary to reset the ADC and MMCM in order to guarantee stable
behavior from them.