utcaApp: remove usage of sentinel parameters.
This is a code cleanup, to avoid having to expose
UDriver::parameter_props to specific drivers.

The cleanup moves all handling of special cases (decoder_controller and
read_only) into the ParamInit class, instead of requiring passing
UDriver static inline members to it. This allowed us to add a
ParamInit::set_dc() function, which is now used by the drivers which
previously had to access parameter_props directly. This also simplifies
the UDriver constructor, since it can now determine the need for p_tmp
more easily, as well as the parameter properties.

It also allowed us to add more sanity checks, in this case to
UDriver::writeGeneral(), to report writes into read-only parameters.