service: attempt a reset when no SDB can be found.
In some conditions, reading from BAR4 will return only 1s, so the SDB
can't even be accessed in order to be decoded. For some of these
situations, resetting the PCIe interface via BAR0 restores BAR4
functionality.

The effects of this reset aren't fully understood, but, since any
further steps to restore communication with the board would include
removing the PCIe slot or restarting the board, any side-effects are
considered benign.